RavenDB-19689 moving items around
