+++ src/Raven.Server/Smuggler/Documents/DatabaseDestination.cs
@@ -6,21 +6,18 @@ using System.IO;
 using System.Runtime.CompilerServices;
 using System.Threading;
 using System.Threading.Tasks;
-using JetBrains.Annotations;
 using Raven.Client.Documents.Attachments;
-using Raven.Client.Documents.Indexes;
 using Raven.Client.Documents.Operations.Attachments;
 using Raven.Client.Documents.Operations.Counters;
 using Raven.Client.Documents.Smuggler;
 using Raven.Client.Exceptions.Documents;
 using Raven.Client.Util;
-using Raven.Server.Config;
 using Raven.Server.Documents;
 using Raven.Server.Documents.Handlers;
-using Raven.Server.Documents.Indexes;
 using Raven.Server.Documents.PeriodicBackup;
 using Raven.Server.Documents.TransactionCommands;
 using Raven.Server.ServerWide.Context;
+using Raven.Server.Smuggler.Documents.Actions;
 using Raven.Server.Smuggler.Documents.Data;
 using Raven.Server.Smuggler.Documents.Processors;
 using Raven.Server.Utils;
@@ -138,12 +135,12 @@ namespace Raven.Server.Smuggler.Documents
 
         public ISubscriptionActions Subscriptions()
         {
-            return new SubscriptionActions(_database);
+            return new DatabaseSubscriptionActions(_database);
         }
 
         public IReplicationHubCertificateActions ReplicationHubCertificates()
         {
-            return new ReplicationHubCertificateActions(_database);
+            return new DatabaseReplicationHubCertificateActions(_database);
         }
 
         public ITimeSeriesActions TimeSeries()
@@ -168,58 +165,6 @@ namespace Raven.Server.Smuggler.Documents
         public IIndexActions Indexes()
         {
             return new DatabaseIndexActions(_database.IndexStore.Create, _database.Time);
-        }
-
-        public class DatabaseIndexActions : IIndexActions
-        {
-            private readonly AbstractIndexCreateController _controller;
-            private readonly SystemTime _time;
-            private readonly AbstractIndexCreateController.IndexBatchScope _batch;
-            private readonly RavenConfiguration _configuration;
-
-            public DatabaseIndexActions([NotNull] AbstractIndexCreateController controller, [NotNull] SystemTime time)
-            {
-                _controller = controller ?? throw new ArgumentNullException(nameof(controller));
-                _time = time ?? throw new ArgumentNullException(nameof(time));
-                _configuration = controller.GetDatabaseConfiguration();
-
-                if (AbstractIndexCreateController.CanUseIndexBatch())
-                    _batch = _controller.CreateIndexBatch();
-            }
-
-            public async ValueTask WriteIndexAsync(IndexDefinitionBaseServerSide indexDefinition, IndexType indexType)
-            {
-                if (_batch != null)
-                {
-                    await _batch.AddIndexAsync(indexDefinition, _source, _time.GetUtcNow(), RaftIdGenerator.DontCareId, _configuration.Indexing.HistoryRevisionsNumber);
-                    await _batch.SaveIfNeeded();
-                    return;
-                }
-
-                await _controller.CreateIndexAsync(indexDefinition, RaftIdGenerator.DontCareId);
-            }
-
-            public async ValueTask WriteIndexAsync(IndexDefinition indexDefinition)
-            {
-                if (_batch != null)
-                {
-                    await _batch.AddIndexAsync(indexDefinition, _source, _time.GetUtcNow(), RaftIdGenerator.DontCareId, _configuration.Indexing.HistoryRevisionsNumber);
-                    await _batch.SaveIfNeeded();
-                    return;
-                }
-
-                await _controller.CreateIndexAsync(indexDefinition, RaftIdGenerator.DontCareId, _source);
-            }
-
-            private const string _source = "Smuggler";
-
-            public async ValueTask DisposeAsync()
-            {
-                if (_batch == null)
-                    return;
-
-                await _batch.SaveAsync();
-            }
         }
 
         public class DuplicateDocsHandler : IDisposable
